+++ src/ch/rabanti/picoxlsx4j/Address.java
@@ -1,6 +1,6 @@
 /*
  * PicoXLSX4j is a small Java library to generate XLSX (Microsoft Excel 2007 or newer) files in an easy and native way
- * Copyright Raphael Stoeckli © 2017
+ * Copyright Raphael Stoeckli © 2018
  * This library is licensed under the MIT License.
  * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
  */
@@ -10,7 +10,7 @@ import static ch.rabanti.picoxlsx4j.Cell.resolveCellAddress;
 import ch.rabanti.picoxlsx4j.exception.RangeException;
 
 /**
- * C class representing a cell address (column and row; used like a simple struct)
+ * C class representing a cell address (no getters and setters to simplify handling)
  * @author Raphael Stoeckli
  */
     public class Address
@@ -28,7 +28,7 @@ import ch.rabanti.picoxlsx4j.exception.RangeException;
         
 // ### C O N S T R U C T O R S ###        
         /**
-         * Constructor with parameters
+         * Constructor with with row and column as arguments
          * @param column Column of the address (zero-based)
          * @param row Row of the address (zero-based)
          */
@@ -37,8 +37,33 @@ import ch.rabanti.picoxlsx4j.exception.RangeException;
             this.Column = column;
             this.Row = row;
         }
+        
+        public Address(String address)
+        {
+         Address a = Cell.resolveCellCoordinate(address);
+         this.Column = a.Column;
+         this.Row = a.Row;
+        }
 
 // ### M E T H O D S ###
+        
+        /**
+         * Compares two addresses whether they are equal
+         * @param o Compares two addresses whether they are equal
+         * @return True if equal
+         */
+        public boolean equals(Address o)
+        {
+            if(this.Row == o.Row && this.Column == o.Column)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
+        }
+        
         /**
          * Gets the address as string
          * @return Address as string