+++ src/ch/rabanti/picoxlsx4j/Cell.java
@@ -1,6 +1,6 @@
 /*
  * PicoXLSX4j is a small Java library to generate XLSX (Microsoft Excel 2007 or newer) files in an easy and native way
- * Copyright Raphael Stoeckli © 2017
+ * Copyright Raphael Stoeckli © 2018
  * This library is licensed under the MIT License.
  * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
  */
@@ -366,7 +366,7 @@ public class Cell implements Comparable<Cell>{
      * @param endRow End row (zero based)
      * @return List of cell addresses
      */
-    public static List<Address> GetCellRange(int startColumn, int startRow, int endColumn, int endRow)
+    public static List<Address> getCellRange(int startColumn, int startRow, int endColumn, int endRow)
     {
         Address start = new Address(startColumn, startRow);
         Address end = new Address(endColumn, endRow);