+++ src/ch/rabanti/picoxlsx4j/lowLevel/LowLevel.java
@@ -1,6 +1,6 @@
 /*
  * PicoXLSX4j is a small Java library to generate XLSX (Microsoft Excel 2007 or newer) files in an easy and native way
- * Copyright Raphael Stoeckli © 2017
+ * Copyright Raphael Stoeckli © 2018
  * This library is licensed under the MIT License.
  * You find a copy of the license in project folder or on: http://opensource.org/licenses/MIT
  */
@@ -48,7 +48,6 @@ import ch.rabanti.picoxlsx4j.style.NumberFormat;
 import ch.rabanti.picoxlsx4j.style.Style;
 import java.io.FileOutputStream;
 import java.io.OutputStream;
-import java.util.stream.Stream;
 
 
 /**