+++ Changelog.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## v2.3.0
+ 
+---
+ 
+Release Date: **13.02.2018**
+- Added most important formulas as static method calls in the class BasicFormulas (round, floor, ceil, min, max, average, median, sum, vlookup)
+- Removed overloaded methods to add cells as type Cell. This can be done now with the overloading of the type object (no code changes necessary)
+- Added new constructors for Address and Range
+- Added demo for the new static formula methods
+- Changed the method names SetColumnWidth (Worksheet) to setColumnWidth and GetCellRange (Cell) to getCellRange, to follow conventions (minor updates of code may necessary)
+- Minor bug fixes
+- Documentation update
+ 
+
 ## v2.2.0
 ---
 Release Date: **10.12.2017**